Set monitor mode in /etc/sysconfig/maldet for OS Family redhat machines.
The service_provider fact on a RedHat Enterprise Linux machine is redhat. But,
on a CentOS machine, it's systemd. I think we still want to manage the
configuration of /etc/sysconfig/maldet, as long as the machine's OS is part of
the RedHat OS family.

Let's just use facts['os']['family'] to determine whether to attempt to manage
the /etc/sysconfig/maldet.